fix(db): bind PostgreSQL parameters at inferred widths
LIMIT/OFFSET and integer-literal comparisons bound i32 where PostgreSQL infers
bigint, and definition and evaluation-manifest advisory locks sent NUL-joined
text. Conformance harnesses now honor the test CA, race tests spawn before
joining, and scenarios that read across namespaces use scratch databases.
A source guard rejects the fixed bind shapes (#1155).

+++ src/db/postgres_attestation.rs
@@ -54,8 +54,8 @@ impl PostgresDb {
                 &[
                     &decision_id,
                     &policy_scope,
-                    &if limit > 0 { limit.min(500) } else { 100 },
-                    &offset.max(0),
+                    &i64::from(if limit > 0 { limit.min(500) } else { 100 }),
+                    &i64::from(offset.max(0)),
                 ],
             )
             .map_err(|error| error.to_string())?
